delete duplicated route for magicLinkRenewal
