+++ back/src/adapters/primary/routers/convention/convention.e2e.test.ts
@@ -13,6 +13,7 @@ import {
   expectHttpResponseToEqual,
   expectToEqual,
   expiredMagicLinkErrorMessage,
+  frontRoutes,
   InclusionConnectedUser,
   stringToMd5,
   TechnicalRoutes,
@@ -33,10 +34,13 @@ import {
   GenerateBackOfficeJwt,
   GenerateConventionJwt,
   GenerateInclusionConnectJwt,
+  makeGenerateJwtES256,
+  makeVerifyJwtES256,
 } from "../../../../domain/auth/jwt";
 import { conventionMissingMessage } from "../../../../domain/convention/entities/Convention";
 import { shortLinkRedirectToLinkWithValidation } from "../../../../utils/e2eTestHelpers";
 import { BasicEventCrawler } from "../../../secondary/core/EventCrawlerImplementations";
+import { AppConfig } from "../../config/appConfig";
 import { InMemoryUnitOfWork } from "../../config/uowConfig";
 
 const peAgency = new AgencyDtoBuilder().withKind("pole-emploi").build();
@@ -65,6 +69,7 @@ describe("convention e2e", () => {
   let inMemoryUow: InMemoryUnitOfWork;
   let eventCrawler: BasicEventCrawler;
   let gateways: InMemoryGateways;
+  let appConfig: AppConfig;
 
   beforeEach(async () => {
     const testApp = await buildTestApp(new AppConfigBuilder().build());
@@ -77,6 +82,7 @@ describe("convention e2e", () => {
       generateBackOfficeJwt,
       generateInclusionConnectJwt,
       inMemoryUow,
+      appConfig,
     } = testApp);
 
     unauthenticatedRequest = createSupertestSharedClient(
@@ -635,6 +641,79 @@ describe("convention e2e", () => {
         },
         status: 404,
       });
+    });
+  });
+
+  describe(`${displayRouteName(
+    conventionMagicLinkRoutes.renewConvention,
+  )} renews a magic link`, () => {
+    it("200 - sends the updated magic link", async () => {
+      const validConvention = new ConventionDtoBuilder().build();
+
+      const agency = AgencyDtoBuilder.create(validConvention.agencyId)
+        .withName("TEST-name")
+        .withAdminEmails(["admin@email.fr"])
+        .withQuestionnaireUrl("TEST-questionnaireUrl")
+        .withSignature("TEST-signature")
+        .build();
+
+      const convention = new ConventionDtoBuilder().build();
+      inMemoryUow.conventionRepository.setConventions([convention]);
+      inMemoryUow.agencyRepository.setAgencies([agency]);
+
+      gateways.timeGateway.setNextDate(new Date());
+
+      generateConventionJwt = makeGenerateJwtES256<"convention">(
+        appConfig.jwtPrivateKey,
+        3600 * 24, // one day
+      );
+      const shortLinkIds = ["shortLink1", "shortLinkg2"];
+      gateways.shortLinkGenerator.addMoreShortLinkIds(shortLinkIds);
+
+      const originalUrl = `${appConfig.immersionFacileBaseUrl}/${frontRoutes.conventionToSign}`;
+
+      const expiredJwt = generateConventionJwt(
+        createConventionMagicLinkPayload({
+          id: validConvention.id,
+          role: "beneficiary",
+          email: validConvention.signatories.beneficiary.email,
+          now: new Date(),
+        }),
+      );
+
+      const response = await unauthenticatedRequest.renewMagicLink({
+        queryParams: {
+          expiredJwt,
+          originalUrl: encodeURIComponent(originalUrl),
+        },
+      });
+
+      expect(response.status).toBe(200);
+
+      await processEventsForEmailToBeSent(eventCrawler);
+
+      const sentEmails = gateways.notification.getSentEmails();
+
+      expect(sentEmails).toHaveLength(1);
+
+      const email = expectEmailOfType(sentEmails[0], "MAGIC_LINK_RENEWAL");
+      expect(email.recipients).toEqual([
+        validConvention.signatories.beneficiary.email,
+      ]);
+
+      const magicLink = await shortLinkRedirectToLinkWithValidation(
+        email.params.magicLink,
+        technicalRoutesClient,
+      );
+
+      const newUrlStart = `${originalUrl}?jwt=`;
+
+      expect(magicLink.startsWith(newUrlStart)).toBeTruthy();
+      const renewedJwt = magicLink.replace(newUrlStart, "");
+      expect(renewedJwt !== expiredJwt).toBeTruthy();
+      expect(
+        makeVerifyJwtES256(appConfig.jwtPublicKey)(renewedJwt),
+      ).toBeDefined();
     });
   });
 });
